fixup! Merge branch 'main' into pr/debug-app-tools
